requirements.txt
django
djangorestframework
django-extra-field
firebase-admin
drf-firebase-authentication
psycopg2
